+++ android/features/shows-grid/src/main/kotlin/com/thomaskioko/tvmaniac/show_grid/ShowGridViewModel.kt
@@ -14,7 +14,7 @@ class ShowGridViewModel @Inject constructor(
     private val stateMachine: GridStateMachine
 ) : ViewModel() {
 
-    val showType: Int = savedStateHandle["showType"]!!
+    val showType: Long = savedStateHandle["showType"]!!
 
     val state: MutableStateFlow<GridState> = MutableStateFlow(LoadingContent)
