+++ shared/domain/following/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/following/api/FollowedShow.kt
@@ -1,8 +1,8 @@
 package com.thomaskioko.tvmaniac.domain.following.api
 
 data class FollowedShow(
-    val traktId: Int = 0,
-    val tmdbId: Int? = 0,
+    val traktId: Long = 0,
+    val tmdbId: Long? = 0,
     val title: String,
     val posterImageUrl: String? = null,
 )
