+++ shared/domain/season-details/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasondetails/api/SeasonDetailsAction.kt
@@ -3,9 +3,9 @@ package com.thomaskioko.tvmaniac.seasondetails.api
 sealed class SeasonDetailsAction
 
 data class LoadSeasonDetails(
-    val showId: Int
+    val showId: Long
 ) : SeasonDetailsAction()
 
 data class ReloadSeasonDetails(
-    val showId: Int
+    val showId: Long
 ): SeasonDetailsAction()
