+++ shared/domain/season-details/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasondetails/api/model/Episode.kt
@@ -1,14 +1,14 @@
 package com.thomaskioko.tvmaniac.seasondetails.api.model
 
 data class Episode(
-    val id: Int,
-    val seasonId: Int,
+    val id: Long,
+    val seasonId: Long,
     val episodeTitle: String,
     val episodeNumberTitle: String,
     val overview: String,
     val imageUrl: String?,
-    val runtime: Int,
-    val voteCount: Int,
+    val runtime: Long,
+    val voteCount: Long,
     val episodeNumber: String,
     val seasonEpisodeNumber: String,
 )
