+++ shared/domain/trailers/api/src/commonMain/kotlin/com.thomaskioko/tvmaniac/domain/trailers/api/TrailersAction.kt
@@ -5,7 +5,7 @@ sealed interface TrailersAction
 object ReloadTrailers: TrailersAction
 
 data class LoadTrailers(
-    val showId: Int,
+    val showId: Long,
     val trailerId: String,
 ) : TrailersAction
