+++ shared/domain/trakt/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/trakt/api/cache/TraktFollowedCache.kt
@@ -16,7 +16,7 @@ interface TraktFollowedCache {
 
     fun observeFollowedShows(): Flow<List<SelectFollowedShows>>
 
-    fun updateShowSyncState(traktId: Int)
+    fun updateShowSyncState(traktId: Long)
 
-    fun removeShow(traktId: Int)
+    fun removeShow(traktId: Long)
 }
